% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bret_cell_markers}
\alias{bret_cell_markers}
\title{Marker Genes for the BRETIGEA method}
\format{
A dataframe with two columns: markers and cell.
\describe{
\item{markers}{Top 1000 marker genes for  astrocyte, endothelial, microglia, neuron, oligodnedrocyte and oligodendrocyte precursor cells.}
\item{cell}{Indicates which of the cell types ( astrocyte, endothelial, microglia, neuron, oligodnedrocyte and oligodendrocyte precursor cells) the marker gene is a marker for.}
}
}
\source{
BRETIGEA
}
\usage{
bret_cell_markers
}
\description{
Marker Genes from the BRETIGEA package
}
\examples{
\dontrun{
 bret_cell_markers
}
}
\keyword{datasets}
