% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bretigea_marker_addition.R
\name{bretigea_marker_addition}
\alias{bretigea_marker_addition}
\title{Runs BRETIGEA using top 1, top 1 & 2, top 1,2,3 ... to top 1...n markers
A function that reruns the BRETIGEA findCells method (modified) on a loop to see
the influence of different combinations of markers in determining the cell-type
proportion estimate.}
\usage{
bretigea_marker_addition(
  count_df,
  bret_cell_markers,
  cell,
  metadata,
  covar,
  pathology_name,
  cell_type_names,
  n
)
}
\arguments{
\item{count_df}{A dataframe with Gene rows and Subject columns.}

\item{bret_cell_markers}{A dataframe with two columns, markers and cell where
markers are genes for cell types and cell indicates the cell type markers
are the gene for.}

\item{cell}{A string indicating which of the cells in the cell type marker list
is being specified to look at.}

\item{metadata}{A dataframe with subjects also in count_df and rows
indicating the subjects id, some covariate, and disease state score or
pathology.}

\item{covar}{A covariate to be taken into account when running linear models
to check the association between the cell type indicated by cell and the
pathology indicated by pathology_names.}

\item{pathology_name}{The pathology associated with the disease in question
for which the association between it and the cell type indicated by cell is
being examined.}

\item{cell_type_names}{The names of all the unique cell types for which
there are markers in bret_cell_markers: unique(bret_cell_markers$cell)}

\item{n}{Specifies how many times BRETIGEA will run BRETIGEA using top 1,
top 1 & 2, top 1,2,3 ... to top 1...n markers}
}
\value{
Returns a graph of the significance of the cell type proportion
specified's association to the pathology indicated upon marker addition
from 0 to n
}
\description{
Runs BRETIGEA using top 1, top 1 & 2, top 1,2,3 ... to top 1...n markers
A function that reruns the BRETIGEA findCells method (modified) on a loop to see
the influence of different combinations of markers in determining the cell-type
proportion estimate.
}
\examples{
# Examples 1:
# Using count_df, bret_cell_markers, metadata datasets available with package

bretigea_marker_addition <- bretigea_marker_addition (
                          count_df = count_df,
                          bret_cell_markers = bret_cell_markers,
                          cell = "mic",
                          metadata = metadata,
                          covar = "Covariate",
                          pathology_name = "DiseasePhenotypeScore",
                          cell_type_names = unique(bret_cell_markers$cell),
                          n= 10)

}
\references{
Mancarci, B. O., Toker, L., Tripathy, S. J., Li, B., Rocco, B., Sibille, E., & Pavlidis, P. (2017).
CrossLaboratory Analysis of Brain Cell Type Transcriptomes with Applications to Interpretation of Bulk
Tissue Data. \emph{eNeuro}, 4(6), ENEURO.0212-17.2017. https://doi.org/10.1523/ENEURO.0212-17.201

Stefan Milton Bache and Hadley Wickham (2014). magrittr: A Forward-Pipe Operator for R.
\emph{R package version 1.5}.https://CRAN.R-project.org/package=magrittr

Wickham et al., (2019). Welcome to the tidyverse.\emph{ Journal of Open Source Software}, 4(43), 1686,
https://doi.org/10.21105/joss.01686
}
