% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bretMarkerEffectOnPathology.R
\name{bretMarkerEffectOnPathology}
\alias{bretMarkerEffectOnPathology}
\title{Runs BRETIGEA using top 1, top 1 & 2, top 1,2,3 ... to top 1...n markers
A function that reruns the BRETIGEA findCells method (modified) on a loop to see
the influence of different combinations of markers in determining the cell-type
proportion estimate.}
\usage{
bretMarkerEffectOnPathology(
  countDf,
  bretCellMarkers,
  cellName,
  metadata,
  covar,
  pathologyName,
  cellTypeNames,
  n
)
}
\arguments{
\item{countDf}{A dataframe with Gene rows and Subject columns.}

\item{bretCellMarkers}{A dataframe with two columns, markers and cell where
markers are genes for cell types and cell indicates the cell type markers
are the gene for.}

\item{cellName}{A string indicating which of the cells in the cell type marker list
is being specified to look at.}

\item{metadata}{A dataframe with subjects also in countDf and rows
indicating the subjects id, some covariate, and disease state score or
pathology.}

\item{covar}{A covariate to be taken into account when running linear models
to check the association between the cell type indicated by cell and the
pathology indicated by pathologyNames.}

\item{pathologyName}{The pathology associated with the disease in question
for which the association between it and the cell type indicated by cell is
being examined.}

\item{cellTypeNames}{The names of all the unique cell types for which
there are markers in bretCellMarkers: unique(bretCellMarkers$cell)}

\item{n}{Specifies how many times BRETIGEA will run BRETIGEA using top 1,
top 1 & 2, top 1,2,3 ... to top 1...n markers}
}
\value{
Returns a graph of the significance of the cell type proportion
specified's association to the pathology indicated upon marker addition
from 0 to n
}
\description{
Runs BRETIGEA using top 1, top 1 & 2, top 1,2,3 ... to top 1...n markers
A function that reruns the BRETIGEA findCells method (modified) on a loop to see
the influence of different combinations of markers in determining the cell-type
proportion estimate.
}
\examples{
# Examples 1:
# Using countDf, bretCellMarkers, metadata datasets available with package

bretMarkerEffectOnPathologyResults <- bretMarkerEffectOnPathology (
                          countDf = countDf,
                          bretCellMarkers = bretCellMarkers,
                          cellName = "mic",
                          metadata = metadata,
                          covar = "Covariate",
                          pathologyName = "DiseasePhenotypeScore",
                          cellTypeNames = unique(bretCellMarkers$cell),
                          n= 10)

}
\references{
Hadley Wickham and Dana Seidel (2020). scales: Scale Functions for Visualization.
R package version 1.1.1. https://CRAN.R-project.org/package=scales

Kirill Müller and Hadley Wickham (2020). tibble: Simple Data Frames. R package
version 3.0.3. https://CRAN.R-project.org/package=tibble

Mancarci, B. O., Toker, L., Tripathy, S. J., Li, B., Rocco, B., Sibille, E., & Pavlidis, P. (2017).
CrossLaboratory Analysis of Brain Cell Type Transcriptomes with Applications to Interpretation of Bulk
Tissue Data. \emph{eNeuro}, 4(6), ENEURO.0212-17.2017. https://doi.org/10.1523/ENEURO.0212-17.201

R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL
https://www.R-project.org/.

Stefan Milton Bache and Hadley Wickham (2014). magrittr: A Forward-Pipe Operator for R.
\emph{R package version 1.5}.https://CRAN.R-project.org/package=magrittr

Wickham et al., (2019). Welcome to the tidyverse.\emph{ Journal of Open Source Software}, 4(43), 1686,
https://doi.org/10.21105/joss.01686
}
