% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAndCompare.R
\name{calcAndCompare}
\alias{calcAndCompare}
\title{Calculates and compares the estimations of cell types derived by the
markerGeneProfile and BRETIGEA findCells method}
\usage{
calcAndCompare(countDf, mgpCellMarkers, bretCellMarkers)
}
\arguments{
\item{countDf}{A dataframe with Gene rows and Subject columns.}

\item{mgpCellMarkers}{A nested A nested list with as many sub-lists of
marker genes as there are for cell types}

\item{bretCellMarkers}{A dataframe with two columns, markers and cell where
markers are genes for cell types and cell indicates the cell type markers
are the gene for.}
}
\value{
Returns a nested list with cell type proportions calculated by the
markerGeneProfile method and the BRETIGEA method and a correlation plot to
compare the two
}
\description{
A function that calculates cell type proportions using two validated methods,
markerGeneProfile and findCells, and then compares the cell type proportions
calculated by each
}
\examples{
# Examples 1:
# Using countDf, bretCellMarkers, mgpCellMarkers datasets available with package

calcAndCompareResults <- calcAndCompare (
                countDf = countDf,
                mgpCellMarkers = mgpCellMarkers,
                bretCellMarkers = bretCellMarkers)

}
\references{
H. Wickham. Reshaping data with the reshape package. Journal of Statistical
Software, 21(12), 2007.

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New
York, 2016.

Kirill Müller and Hadley Wickham (2020). tibble: Simple Data Frames. R package
version 3.0.3. https://CRAN.R-project.org/package=tibble

Mancarci, B. O., Toker, L., Tripathy, S. J., Li, B., Rocco, B., Sibille, E., & Pavlidis, P. (2017).
CrossLaboratory Analysis of Brain Cell Type Transcriptomes with Applications to Interpretation of Bulk
Tissue Data. \emph{eNeuro}, 4(6), ENEURO.0212-17.2017. https://doi.org/10.1523/ENEURO.0212-17.201

McKenzie, A.T., Wang, M., Hauberg, M.E. et al. Brain Cell Type Specific Gene Expression and Coexpression Network Architectures.
\emph{Sci Rep} 8, 8868 (2018). https://doi.org/10.1038/s41598-018-27293-5

R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.

Wickham et al., (2019). Welcome to the tidyverse.\emph{ Journal of Open Source Software}, 4(43), 1686,
https://doi.org/10.21105/joss.01686
}
