% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_df}
\alias{count_df}
\title{Bulk-tissue RNAseq data counts}
\format{
A dataframe with 395 rows of genes and 378 subjects.
\describe{
\item{Gene}{List of genes, some of which may be marker genes}
\item{Subjects}{378 subjects}
}
}
\source{
BRETIGEA
}
\usage{
count_df
}
\description{
Taken from the BRETIGEA package (originally aba_marker_expression)
}
\examples{
\dontrun{
 count_df
}
}
\keyword{datasets}
