% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCellsMod.R
\name{findCellsMod}
\alias{findCellsMod}
\title{Modified BRETIGEA findCells() function}
\usage{
findCellsMod(inputMat, markers, nMarker, method, scale)
}
\arguments{
\item{inputMat}{Numeric gene expression data frame or matrix, with
rownames corresponding to gene names, some of which are marker genes,
and columns corresponding to samples.}

\item{markers}{Data frame with marker genes in one column (named "marker")
and the cell type that that gene symbol corresponds to in another column
(named "cell").}

\item{nMarker}{The number of marker genes (that are present in your expression
data set) to use in estimating the surrogate cell type proportion variable
for each cell type.}

\item{method}{To estimate the cell type proportions, can either use PCA or SVD.}

\item{scale}{Whether or not to scale the gene expression data from each
marker gene prior to using it as an input for dimension reduction.}
}
\value{
Returns a list of sample-by-cell type matrix of estimated cell
type proportion variables and the markers used
}
\description{
A function that runs BRETIGEA findCells() but returns the used markers
}
\examples{
# Examples 1:
# Using bretCellMarkers, metadata datasets available with package

rownames(countDf) <- countDf$Gene
countDf <- countDf[,-1]
findCellsMod <- findCellsMod(
               inputMat = countDf,
               markers = bretCellMarkers,
               nMarker = 10,
               method = "SVD",
               scale = TRUE)
findCellsMod$markerList
findCellsMod$SPVS

}
\references{
#' Chikina M, Zaslavsky E, Sealfon SC. CellCODE: a robust latent variable approach to differential expression analysis for heterogeneous cell populations. \emph{Bioinformatics}
. 2015;31(10):1584-91.

Mancarci, B. O., Toker, L., Tripathy, S. J., Li, B., Rocco, B., Sibille, E., & Pavlidis, P. (2017).
CrossLaboratory Analysis of Brain Cell Type Transcriptomes with Applications to Interpretation of Bulk
Tissue Data. \emph{eNeuro}, 4(6), ENEURO.0212-17.2017. https://doi.org/10.1523/ENEURO.0212-17.201

R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.
}
