% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_used.R
\name{genes_used}
\alias{genes_used}
\title{Returns genes used in Marker Gene Profile and BRETIGEA method}
\usage{
genes_used(count_df, mgp_cell_markers, bret_cell_markers)
}
\arguments{
\item{count_df}{A dataframe with Gene rows and Subject columns.}

\item{mgp_cell_markers}{A nested A nested list with as many sub-lists of
marker genes as there are for cell types}

\item{bret_cell_markers}{A dataframe with two columns, markers and cell where
markers are genes for cell types and cell indicates the cell type markers
are the gene for.}
}
\value{
Returns a nested list with mgp_markers and bret_markers which are each
}
\description{
A function that returns the genes used in the markerGenProfile estimation method
and the BRETIGEA method
}
\examples{
# Examples 1:
# Using count_df, bret_cell_markers, mgp_cell_markers datasets available with package

genes_used <- genes_used (
                count_df = count_df,
                mgp_cell_markers = mgp_cell_markers,
                bret_cell_markers = bret_cell_markers)

}
\references{
Mancarci, B. O., Toker, L., Tripathy, S. J., Li, B., Rocco, B., Sibille, E., & Pavlidis, P. (2017).
CrossLaboratory Analysis of Brain Cell Type Transcriptomes with Applications to Interpretation of Bulk
Tissue Data. \emph{eNeuro}, 4(6), ENEURO.0212-17.2017.https://doi.org/10.1523/ENEURO.0212-17.201

McKenzie, A.T., Wang, M., Hauberg, M.E. et al. Brain Cell Type Specific Gene Expression and Coexpression Network Architectures.
\emph{Sci Rep} 8, 8868 (2018). https://doi.org/10.1038/s41598-018-27293-5
}
