% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesUsed.R
\name{genesUsed}
\alias{genesUsed}
\title{Returns genes used in Marker Gene Profile and BRETIGEA method}
\usage{
genesUsed(countDf, mgpCellMarkers, bretCellMarkers)
}
\arguments{
\item{countDf}{A dataframe with Gene rows and Subject columns.}

\item{mgpCellMarkers}{A nested A nested list with as many sub-lists of
marker genes as there are for cell types}

\item{bretCellMarkers}{A dataframe with two columns, markers and cell where
markers are genes for cell types and cell indicates the cell type markers
are the gene for.}
}
\value{
Returns a nested list with mgpMarkers and bretMarkers which are each
lists of markers used for each method
}
\description{
A function that returns the genes used in the markerGeneProfile estimation method
and the BRETIGEA method
}
\examples{
# Examples 1:
# Using countDf, bretCellMarkers, mgpCellMarkers datasets available with package

genesUsedResults <- genesUsed (
                countDf = countDf,
                mgpCellMarkers = mgpCellMarkers,
                bretCellMarkers = bretCellMarkers)

}
\references{
Mancarci, B. O., Toker, L., Tripathy, S. J., Li, B., Rocco, B., Sibille, E., & Pavlidis, P. (2017).
CrossLaboratory Analysis of Brain Cell Type Transcriptomes with Applications to Interpretation of Bulk
Tissue Data. \emph{eNeuro}, 4(6), ENEURO.0212-17.2017.https://doi.org/10.1523/ENEURO.0212-17.201

McKenzie, A.T., Wang, M., Hauberg, M.E. et al. Brain Cell Type Specific Gene Expression and Coexpression Network Architectures.
\emph{Sci Rep} 8, 8868 (2018). https://doi.org/10.1038/s41598-018-27293-5

Stefan Milton Bache and Hadley Wickham (2014). magrittr: A Forward-Pipe Operator
for R. R package version 1.5. https://CRAN.R-project.org/package=magrittr
}
