% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bretigea_marker_addition.R
\name{markers_pathology}
\alias{markers_pathology}
\title{Determines associations between cell type estimates and pathology}
\usage{
markers_pathology(markerMat, metadata, covar, pathology_name, cell_type_names)
}
\arguments{
\item{markerMat}{The $SPVS of the return of find_cells_mod once converted to a
dataframe with the first column being Sample corresponding to the metadata
Sample id. (Dataframe with Sample column and then estimate of cell type
proportion calculated by BRETIGEA findCells for each cell type in
cell_type_names)}

\item{metadata}{A dataframe with subjects also in count_df and rows
indicating the subjects id, some covariate, and disease state score or
pathology.}

\item{covar}{A covariate to be taken into account when running linear models
to check the association between the cell type indicated by cell and the
pathology indicated by pathology_names.}

\item{pathology_name}{The pathology associated with the disease in question
for which the association between it and the cell type indicated by cell is
being examined.}

\item{cell_type_names}{The names of all the unique cell types for which
there are markers in bret_cell_markers: unique(bret_cell_markers$cell)}
}
\value{
Returns a matrix array ready to be formatted by
}
\description{
A function that runs a linear model on the cell-type
proportion estimates by BRETIGEA.
}
\examples{
# Examples 1:
# Using bret_cell_markers, metadata datasets available with package

rownames(count_df) <- count_df$Gene
count_df <- count_df[,-1]
bretigeaMarkers <- find_cells_mod(
               inputMat = count_df,
               markers = bret_cell_markers,
               nMarker = 10,
               method = "SVD",
               scale = TRUE)
markerMat <- as.data.frame(bretigeaMarkers$SPVS)
markerMat <- tibble::rownames_to_column(markerMat, var = 'Sample')
markers_pathology <- markers_pathology(
                      markerMat = markerMat,
                      metadata = metadata,
                      covar = "Covariate",
                      pathology_name = "DiseasePhenotypeScore",
                      cell_type_names = unique(bret_cell_markers$cell))

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The New S Language}. Wadsworth & Brooks/Cole.

Chambers, J. M. and Hastie, T. J. (1992) \emph{Statistical Models in S}, Wadsworth & Brooks/Cole.

R Core Team (2020). R: A language and environment for statistical computing.
R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.

Wilkinson, G. N. and Rogers, C. E. (1973). Symbolic descriptions of factorial models for analysis of variance. \emph{Applied Statistics}, 22, 392–399. doi: 10.2307/2346786.
}
