% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bretMarkerEffectOnPathology.R
\name{markersPathology}
\alias{markersPathology}
\title{Determines associations between cell type estimates and pathology}
\usage{
markersPathology(markerMat, metadata, covar, pathologyName, cellTypeNames)
}
\arguments{
\item{markerMat}{The $SPVS of the return of findCellsMod once converted to a
dataframe with the first column being Sample corresponding to the metadata
Sample id. (Dataframe with Sample column and then estimate of cell type
proportion calculated by BRETIGEA findCells for each cell type in
cellTypeNames)}

\item{metadata}{A dataframe with subjects also in countDf and rows
indicating the subjects id, some covariate, and disease state score or
pathology.}

\item{covar}{A covariate to be taken into account when running linear models
to check the association between the cell type indicated by cell and the
pathology indicated by pathologyNames.}

\item{pathologyName}{The pathology associated with the disease in question
for which the association between it and the cell type indicated by cell is
being examined.}

\item{cellTypeNames}{The names of all the unique cell types for which
there are markers in bretCellMarkers: unique(bretCellMarkers$cell)}
}
\value{
Returns a matrix array ready to be formatted by bretigeaMarkersAddition
}
\description{
A function that runs a linear model on the cell-type
proportion estimates by BRETIGEA.
}
\examples{
# Examples 1:
# Using bretCellMarkers, metadata datasets available with package

rownames(countDf) <- countDf$Gene
countDf <- countDf[,-1]
bretigeaMarkers <- findCellsMod(
               inputMat = countDf,
               markers = bretCellMarkers,
               nMarker = 10,
               method = "SVD",
               scale = TRUE)
markerMat <- as.data.frame(bretigeaMarkers$SPVS)
markerMat <- tibble::rownames_to_column(markerMat, var = 'Sample')
markersPathologyResults <- markersPathology(
                      markerMat = markerMat,
                      metadata = metadata,
                      covar = "Covariate",
                      pathologyName = "DiseasePhenotypeScore",
                      cellTypeNames = unique(bretCellMarkers$cell))

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The New S Language}. Wadsworth & Brooks/Cole.

Chambers, J. M. and Hastie, T. J. (1992) \emph{Statistical Models in S}, Wadsworth & Brooks/Cole.

R Core Team (2020). R: A language and environment for statistical computing.
R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.

Wilkinson, G. N. and Rogers, C. E. (1973). Symbolic descriptions of factorial models for analysis of variance. \emph{Applied Statistics}, 22, 392–399. doi: 10.2307/2346786.
}
