% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata for subjects corresponding to count_df subjects}
\format{
A dataframe with 345 subjects also in count_df and 3 rows indicating the subjects id (Sample), some covariate (covariate) and disease state score (DiseasePhenotypeScore)
\describe{
 \item{Sample}{The subjects id which corresponds to count_df}
 \item{Covariate}{Some confounding factor that must be accounted for when finding signals between cell-type proportions and a disease phenotype (could be batch etc.)}
 \item{DiseasePhenotypeScore}{Some indication of disease state, could be a score of combined AD pathology variables (CERAD score, Braak pathology stage etc.)}
}
}
\source{
BRETIGEA
}
\usage{
metadata
}
\description{
Taken from the BRETIGEA package, modified to have "Covariate" and "DiseasePhenotypeScore" columns from previous variables (names are not meaningful, just used to show relationships between cell type proportions and potentially a disease phenotype state, covarying for some confounding effect)
}
\examples{
\dontrun{
 metadata
}
}
\keyword{datasets}
