% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgp_cell_markers}
\alias{mgp_cell_markers}
\title{Marker Genes for the Marker Gene Profile method}
\format{
A nested list with 6 sub-lists of marker genes for brain cells: astrocyte, endothelial, microglia, neuron, oligodnedrocyte and oligodendrocyte precursor cells:
\describe{
 \item{ast}{Top 50 astrocyte marker genes.}
 \item{end}{Top 50 astrocyte marker genes.}
 \item{mic}{Top 50 microglia marker genes.}
 \item{neu}{Top 50 neuron marker genes.}
 \item{oli}{Top 50 oligodendrocyte marker genes.}
 \item{opc}{Top 50 oligodendrocyte precursor cell marker genes.}
}
}
\source{
BRETIGEA
}
\usage{
mgp_cell_markers
}
\description{
Marker Genes from the BRETIGEA packaged reformated for use in the Marker Gene Profile method
}
\examples{
\dontrun{
 mgp_cell_markers
}
}
\keyword{datasets}
