% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgpQCMetrics.R
\name{mgpQCMetrics}
\alias{mgpQCMetrics}
\title{Returns genes used, removedMarkerRatios and proportion of variance explained
by the first PC per cell type in the Marker Gene Profile method}
\usage{
mgpQCMetrics(countDf, mgpCellMarkers)
}
\arguments{
\item{countDf}{A dataframe with a Gene row in HGNC symbols and then Subject columns.}

\item{mgpCellMarkers}{A nested list with as many sub-lists of
marker genes as there are for cell types}
}
\value{
Returns a dataframe with the cell type, markersUsed (list of marker genes used
per cell type), removedMarkerRatios (list of removed marker
ratios per cell type) and percentVariancePC1 (list of variance explained
by the first PC per cell type)
}
\description{
A function that returns a df of the genes used, removedMarkerRatios and proportion
of variance explained by the first PC in the markerGeneProfile estimation method
}
\examples{
# Examples 1:
# Using countDf and mgpCellMarkers datasets available with package

mgpQCMetricsResults <- mgpQCMetrics(countDf, mgpCellMarkers)

}
\references{
Mancarci, B. O., Toker, L., Tripathy, S. J., Li, B., Rocco, B., Sibille, E., & Pavlidis, P. (2017).
CrossLaboratory Analysis of Brain Cell Type Transcriptomes with Applications to Interpretation of Bulk
Tissue Data. \emph{eNeuro}, 4(6), ENEURO.0212-17.2017.https://doi.org/10.1523/ENEURO.0212-17.201

Stefan Milton Bache and Hadley Wickham (2014). magrittr: A Forward-Pipe Operator
for R. R package version 1.5. https://CRAN.R-project.org/package=magrittr
}
