% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCellTyPETool.R
\name{runCellTyPETool}
\alias{runCellTyPETool}
\title{Launch Shiny App for CellTyPETool}
\usage{
runCellTyPETool()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for CellTyPETool
The purpose of this app is
}
\examples{
\dontrun{

CellTyPETool::runCellTyPETool()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
