% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_v_phenotype.R
\name{estimates_v_phenotype}
\alias{estimates_v_phenotype}
\title{Calculates and compares the estimations of cell types derived by the
markerGeneProfile and BRETIGEA findCells method}
\usage{
estimates_v_phenotype(
  estimates,
  metadata,
  cell_type_names,
  covar,
  pathology_name
)
}
\arguments{
\item{estimates}{The estimates of cell type proportions returned by calc_and_compare()
either markerGeneProfile derived or BRETIGEA derived}

\item{metadata}{A dataframe with subjects also in count_df and rows
indicating the subjects id, some covariate, and disease state score or
pathology.}

\item{cell_type_names}{The names of all the unique cell types for which
there are markers in bret_cell_markers: unique(bret_cell_markers$cell)}

\item{covar}{A covariate to be taken into account when running linear models
to check the association between the cell type indicated by cell and the
pathology indicated by pathology_name.}

\item{pathology_name}{The pathology associated with the disease in question
for which the association between it and the cell type indicated by cell is
being examined.}
}
\value{
Returns a volcano plot showing the significance of assocations between
each cell type proportion derived and the pathology in question
}
\description{
A function that calculates cell type proportions using two validated methods,
markerGeneProfile and findCells, and then compares the cell type proportions
calculated by each
}
\examples{
# Examples 1:
# Using count_df, bret_cell_markers, mgp_cell_markers datasets available with package

calc_and_compare <- calc_and_compare (
                count_df = count_df,
                mgp_cell_markers = mgp_cell_markers,
                bret_cell_markers = bret_cell_markers)

estimates_v_phenotype <- estimates_v_phenotype(
                estimates = calc_and_compare$bret,
                metadata = metadata,
                cell_type_names = unique(bret_cell_markers$cell),
                covar = "Covariate",
                pathology = "DiseasePhenotypeScore")

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New
York, 2016.

Kamil Slowikowski (2020). ggrepel: Automatically Position Non-Overlapping Text
Labels with 'ggplot2'. R package version 0.8.2.
https://CRAN.R-project.org/package=ggrepel

Mancarci, B. O., Toker, L., Tripathy, S. J., Li, B., Rocco, B., Sibille, E., & Pavlidis, P. (2017).
CrossLaboratory Analysis of Brain Cell Type Transcriptomes with Applications to Interpretation of Bulk
Tissue Data. \emph{eNeuro}, 4(6), ENEURO.0212-17.2017. https://doi.org/10.1523/ENEURO.0212-17.201

McKenzie, A.T., Wang, M., Hauberg, M.E. et al. Brain Cell Type Specific Gene Expression and Coexpression Network Architectures.
\emph{Sci Rep} 8, 8868 (2018). https://doi.org/10.1038/s41598-018-27293-5

R Core Team (2020). R: A language and environment for statistical computing. R
Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.

Wickham et al., (2019). Welcome to the tidyverse.\emph{ Journal of Open Source Software}, 4(43), 1686,
https://doi.org/10.21105/joss.01686

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York. ISBN 978-3-319-24277-4, https://ggplot2.tidyverse.org.
}
